InsightMate/InsightMateApp/Resources/py/requirements.txt
flask
openai
google-auth
google-auth-oauthlib
google-api-python-client
requests
msal
